First we pin all the legacy behavior in place. We make no attempt to understand it. We just want to ensure that it doesn't change.
Approvals Tests are great for this. We just run through a bunch of iterations, build up a log file of all the intermediate and final states, and then approve the log. Now any change will show up as a difference in the log - and the diff will usually make it pretty obvious what we screwed up.

I'm using the QuietReporter since I use NCrunch. I don't want it to pop a diff window every time that I type a character.