I was getting a lot of helpers, so wanted to pull them out. Also, I noticed that they were all statics. That means they're probably on the wrong class.
They all take a single argument (the item), so should be on that class. We have discovered some of the missing domain methods. Item is a severe case of primitive obsession. We could fix that by moving these methods.

Unfortunately, the drag on will eat us if we touch the class. So I'm going to make them into extension methods.

I select one of them and tell R# to move it. I then type in the name of the new class to create, tell it to make them public, and poof! It creates a class and moves everything over. It even figured out that the class should be a static class.